# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Product API",
        "description": "API для управления товарами, арендой и покупками в системе Laravel Product",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Локальный сервер разработки"
        }
    ],
    "paths": {
...
        "/api/products/{productId}/status": {
            "get": {
                "tags": [
                    "Product Status"
                ],
                "summary": "Проверить статус товара для пользователя",
                "description": "Проверяет статус конкретного товара для текущего авторизованного пользователя. Возвращает информацию о владении товаром.",
                "operationId": "checkUserProductStatus",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "Идентификатор товара",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статус товара получен успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductStatusResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Товар не найден в ваших владениях",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/status/{uniqueCode}": {
            "get": {
                "tags": [
                    "Product Status"
                ],
                "summary": "Проверить статус товара по уникальному коду",
                "description": "Проверяет статус товара по его уникальному коду. Не требует авторизации.",
                "operationId": "checkStatusByUniqueCode",
                "parameters": [
                    {
                        "name": "uniqueCode",
                        "in": "path",
                        "description": "Уникальный код товара",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "PROD-001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статус товара получен успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductStatusResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Товар с указанным кодом не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/ownerships": {
            "get": {
                "tags": [
                    "Product Status"
                ],
                "summary": "Получить все владения пользователя",
                "description": "Возвращает список всех товаров, которыми владеет текущий пользователь (купленные и арендованные). Требуется авторизация.",
                "operationId": "getUserOwnerships",
                "responses": {
                    "200": {
                        "description": "Список владений получен успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserOwnershipsResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/purchase": {
            "post": {
                "tags": [
                    "Purchase"
                ],
                "summary": "Купить товар",
                "description": "Позволяет пользователю купить товар. После покупки товар становится собственностью пользователя навсегда. Требуется авторизация.",
                "operationId": "purchaseProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PurchaseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Товар успешно куплен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OwnershipResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации или бизнес-логики (недостаточно средств, товар уже куплен)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Товар или пользователь не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "title": "Ответ с ошибкой",
                "description": "Стандартный ответ API при ошибке",
                "properties": {
                    "success": {
                        "description": "Статус успешности операции",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Сообщение об ошибке",
                        "type": "string",
                        "example": "Товар не найден"
                    },
                    "errors": {
                        "description": "Детали ошибок (опционально)",
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ProductStatus": {
                "title": "Статус товара",
                "properties": {
                    "id": {
                        "description": "Идентификатор владения",
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "description": "Идентификатор товара",
                        "type": "integer",
                        "example": 1
                    },
                    "product_name": {
                        "description": "Название товара",
                        "type": "string",
                        "example": "Ноутбук Dell XPS 13"
                    },
                    "unique_code": {
                        "description": "Уникальный код товара",
                        "type": "string",
                        "example": "PROD-001"
                    },
                    "type": {
                        "description": "Тип владения",
                        "type": "string",
                        "enum": [
                            "rent",
                            "purchase"
                        ],
                        "example": "rent"
                    },
                    "status": {
                        "description": "Статус владения",
                        "type": "string",
                        "enum": [
                            "active",
                            "expired"
                        ],
                        "example": "active"
                    },
                    "expires_at": {
                        "description": "Дата окончания аренды (только для аренды)",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T12:00:00Z",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "ProductStatusResponse": {
                "title": "Ответ со статусом товара",
                "properties": {
                    "success": {
                        "description": "Статус успешности операции",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "Сообщение о результате операции",
                        "type": "string",
                        "example": "Статус товара получен успешно"
                    },
                    "data": {
                        "$ref": "#/components/schemas/ProductStatus"
                    }
                },
                "type": "object"
            },
            "UserOwnershipsResponse": {
                "title": "Ответ со списком владений пользователя",
                "properties": {
                    "success": {
                        "description": "Статус успешности операции",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "Сообщение о результате операции",
                        "type": "string",
                        "example": "Список ваших товаров получен успешно"
                    },
                    "data": {
                        "description": "Список владений пользователя",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductStatus"
                        }
                    }
                },
                "type": "object"
            },
            "PurchaseRequest": {
                "title": "Запрос на покупку товара",
                "required": [
                    "product_id",
                    "unique_code"
                ],
                "properties": {
                    "product_id": {
                        "description": "Идентификатор товара для покупки",
                        "type": "integer",
                        "example": 1
                    },
                    "unique_code": {
                        "description": "Уникальный код товара",
                        "type": "string",
                        "example": "PROD-001"
                    }
                },
                "type": "object"
            },
...
            "Ownership": {
                "title": "Владение товаром",
                "properties": {
                    "id": {
                        "description": "Идентификатор владения",
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "description": "Идентификатор товара",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "Идентификатор пользователя",
                        "type": "integer",
                        "example": 1
                    },
                    "unique_code": {
                        "description": "Уникальный код товара",
                        "type": "string",
                        "example": "PROD-001"
                    },
                    "type": {
                        "description": "Тип владения",
                        "type": "string",
                        "enum": [
                            "rent",
                            "purchase"
                        ],
                        "example": "rent"
                    },
                    "expires_at": {
                        "description": "Дата окончания аренды (только для аренды)",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T12:00:00Z"
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Дата последнего обновления",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "OwnershipResponse": {
                "title": "Ответ с данными владения",
                "properties": {
                    "success": {
                        "description": "Статус успешности операции",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "Сообщение о результате операции",
                        "type": "string",
                        "example": "Товар успешно арендован"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Ownership"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Аутентификация через Laravel Sanctum токены",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}